Make it easier to create a customized user-exception-widget.
